pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prime-zeta-research"
version = "0.1.0"
description = "Research on connection between prime numbers and Riemann zeta function zeros"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "sympy>=1.12.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "PyWavelets>=1.4.0",
    "mpmath>=1.3.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "docling>=2.72.0",
    "manim>=0.19.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
